requirements.txt
fastapi
uvicorn
opencv-python
numpy
matplotlib
ultralytics
google-generativeai
shutil